requirements.txt
python-dotenv
requests
pandas
numpy
markdown
